lib: nrf_modem: update nrf91 offloaded net device API
The API for an offloaded net device has changed from net_if_api
to offloaded_if_api, which introduced an `enable` function pointer.

Give that the API structure was passed to a macro, no parameter
checking guaranteed that the type was correct. Eventually
the networking subsystem would call enable(), which being
uninitialized resulted in a MPU fault.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>